+++ crossdock/thrift/echo/yarpc/echoclient/client.go
@@ -24,14 +24,14 @@
 package echoclient
 
 import (
-	"github.com/thriftrw/thriftrw-go/protocol"
+	yarpc "github.com/yarpc/yarpc-go"
 	"golang.org/x/net/context"
-	"github.com/yarpc/yarpc-go"
+	"go.uber.org/thriftrw/protocol"
+	"go.uber.org/thriftrw/wire"
 	"github.com/yarpc/yarpc-go/crossdock/thrift/echo"
 	"github.com/yarpc/yarpc-go/transport"
 	"github.com/yarpc/yarpc-go/encoding/thrift"
 	echo2 "github.com/yarpc/yarpc-go/crossdock/thrift/echo/service/echo"
-	"github.com/thriftrw/thriftrw-go/wire"
 )
 
 // Interface is a client for the Echo service.