+++ crossdock/thrift/gauntlet/service/thrifttest/testvoid.go
@@ -25,7 +25,7 @@ package thrifttest
 
 import (
 	"fmt"
-	"github.com/thriftrw/thriftrw-go/wire"
+	"go.uber.org/thriftrw/wire"
 	"strings"
 )