+++ crossdock/thrift/gauntlet/yarpc/secondserviceserver/server.go
@@ -24,12 +24,12 @@
 package secondserviceserver
 
 import (
-	"github.com/thriftrw/thriftrw-go/protocol"
+	yarpc "github.com/yarpc/yarpc-go"
 	"golang.org/x/net/context"
-	"github.com/yarpc/yarpc-go"
+	"go.uber.org/thriftrw/protocol"
+	"go.uber.org/thriftrw/wire"
 	"github.com/yarpc/yarpc-go/crossdock/thrift/gauntlet/service/secondservice"
 	"github.com/yarpc/yarpc-go/encoding/thrift"
-	"github.com/thriftrw/thriftrw-go/wire"
 )
 
 // Interface is the server-side interface for the SecondService service.