+++ crossdock/thrift/gauntlet/yarpc/thrifttestserver/server.go
@@ -24,13 +24,13 @@
 package thrifttestserver
 
 import (
-	"github.com/thriftrw/thriftrw-go/protocol"
+	yarpc "github.com/yarpc/yarpc-go"
 	"golang.org/x/net/context"
-	"github.com/yarpc/yarpc-go"
+	"go.uber.org/thriftrw/protocol"
+	"go.uber.org/thriftrw/wire"
 	"github.com/yarpc/yarpc-go/crossdock/thrift/gauntlet"
 	"github.com/yarpc/yarpc-go/encoding/thrift"
 	"github.com/yarpc/yarpc-go/crossdock/thrift/gauntlet/service/thrifttest"
-	"github.com/thriftrw/thriftrw-go/wire"
 )
 
 // Interface is the server-side interface for the ThriftTest service.