+++ examples/thrift/hello/thrift/hello/types.go
@@ -26,7 +26,7 @@ package hello
 import (
 	"errors"
 	"fmt"
-	"github.com/thriftrw/thriftrw-go/wire"
+	"go.uber.org/thriftrw/wire"
 	"strings"
 )