+++ examples/thrift/hello/thrift/hello/yarpc/helloclient/client.go
@@ -24,14 +24,14 @@
 package helloclient
 
 import (
-	"github.com/thriftrw/thriftrw-go/protocol"
+	yarpc "github.com/yarpc/yarpc-go"
 	"golang.org/x/net/context"
-	"github.com/yarpc/yarpc-go"
+	"go.uber.org/thriftrw/protocol"
+	"go.uber.org/thriftrw/wire"
 	"github.com/yarpc/yarpc-go/encoding/thrift"
 	"github.com/yarpc/yarpc-go/transport"
 	"github.com/yarpc/yarpc-go/examples/thrift/hello/thrift/hello"
 	hello2 "github.com/yarpc/yarpc-go/examples/thrift/hello/thrift/hello/service/hello"
-	"github.com/thriftrw/thriftrw-go/wire"
 )
 
 // Interface is a client for the Hello service.