+++ examples/thrift/hello/thrift/hello/yarpc/helloserver/server.go
@@ -24,13 +24,13 @@
 package helloserver
 
 import (
-	"github.com/thriftrw/thriftrw-go/protocol"
+	yarpc "github.com/yarpc/yarpc-go"
 	"golang.org/x/net/context"
-	"github.com/yarpc/yarpc-go"
+	"go.uber.org/thriftrw/protocol"
+	"go.uber.org/thriftrw/wire"
 	"github.com/yarpc/yarpc-go/encoding/thrift"
 	"github.com/yarpc/yarpc-go/examples/thrift/hello/thrift/hello"
 	hello2 "github.com/yarpc/yarpc-go/examples/thrift/hello/thrift/hello/service/hello"
-	"github.com/thriftrw/thriftrw-go/wire"
 )
 
 // Interface is the server-side interface for the Hello service.