+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Releases
 v0.3.0 (unreleased)
 -------------------
 
+-   **Breaking**: Switch to `go.uber.org/thriftrw ~0.3` from
+    `github.com/thriftrw/thriftrw-go ~0.2`.
 -   Update opentracing-go to `>= 1, < 2`.